+++ tests/test_extra/test_custom.py
@@ -47,6 +47,20 @@ class TestCustomSelectors(util.TestCase):
             flags=util.HTML
         )
 
+    def test_custom_selectors_exotic(self):
+        """Test custom selectors."""
+
+        custom_selectors = sv.Custom()
+        custom_selectors.append(":--", "h1, h2, h3, h4, h5, h6")
+
+        self.assert_selector(
+            self.MARKUP,
+            ':--',
+            ['1', '2'],
+            custom=custom_selectors,
+            flags=util.HTML
+        )
+
     def test_custom_dependency(self):
         """Test custom selector dependency on other custom selectors."""