+++ docs/src/markdown/about/changelog.md
@@ -5,7 +5,7 @@
 - **NEW**: Add custom selector support. (#92)
 - **FIX**: Small tweak to CSS identifier pattern to ensure it matches the CSS specification exactly. Specifically, you
 can't have an identifier of only `-`.
-- **FIX**: CSS string patterns should allow escaping newlines.
+- **FIX**: CSS string patterns should allow escaping newlines to span strings across multiple lines.
 - **FIX**: Newline regular expression for CSS newlines should treat `\r\n` as a single character, especially in cases
 such as string escapes: `\\\r\n`.
 - **FIX**: Allow `--` as a valid identifier or identifier start.